update floor dto float

+++ Data.ViewModels/MstFloorDto.cs
@@ -10,10 +10,10 @@ namespace Data.ViewModels
         public Guid BuildingId { get; set; }
         public string Name { get; set; }
         public string FloorImage { get; set; }
-        public long PixelX { get; set; }
-        public long PixelY { get; set; }
-        public long FloorX { get; set; }
-        public long FloorY { get; set; }
+        public float PixelX { get; set; }
+        public float PixelY { get; set; }
+        public float FloorX { get; set; }
+        public float FloorY { get; set; }
         public float MeterPerPx { get; set; }
         public long EngineFloorId { get; set; }
         public string CreatedBy { get; set; }
@@ -29,10 +29,10 @@ namespace Data.ViewModels
         public Guid BuildingId { get; set; }
         public string Name { get; set; }
         public IFormFile FloorImage { get; set; }
-        public long PixelX { get; set; }
-        public long PixelY { get; set; }
-        public long FloorX { get; set; }
-        public long FloorY { get; set; }
+        public float PixelX { get; set; }
+        public float PixelY { get; set; }
+        public float FloorX { get; set; }
+        public float FloorY { get; set; }
         public float MeterPerPx { get; set; }
         public long EngineFloorId { get; set; }
     }
@@ -42,10 +42,10 @@ namespace Data.ViewModels
         public Guid BuildingId { get; set; }
         public string Name { get; set; }
         public IFormFile FloorImage { get; set; }
-        public long PixelX { get; set; }
-        public long PixelY { get; set; }
-        public long FloorX { get; set; }
-        public long FloorY { get; set; }
+        public float PixelX { get; set; }
+        public float PixelY { get; set; }
+        public float FloorX { get; set; }
+        public float FloorY { get; set; }
         public float MeterPerPx { get; set; }
         public long EngineFloorId { get; set; }
     }
@@ -54,10 +54,10 @@ namespace Data.ViewModels
         public string BuildingId { get; set; } 
         public string Name { get; set; }
         public string FloorImage { get; set; } = "";
-        public long PixelX { get; set; }
-        public long PixelY { get; set; }
-        public long FloorX { get; set; }
-        public long FloorY { get; set; }
+        public float PixelX { get; set; }
+        public float PixelY { get; set; }
+        public float FloorX { get; set; }
+        public float FloorY { get; set; }
         public float MeterPerPx { get; set; }
         public long EngineFloorId { get; set; }
         public string CreatedBy { get; set; }
